+++ changelog.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Table of Content
 - Added Blog Table of Content
 - Added Dropdown of Header
+- Added contact sub pages
 
 ### Removed